Add achievements screen and fix paging
Adds a full achievements flow with domain models, API wiring, UI, and bottom sheet support, plus profile/stats navigation updates. Also aligns several paginated screens to zero-based paging, maps new profile follow codes and achievement types from DTOs, and enables cleartext traffic for Android.

+++ designSystem/src/commonMain/kotlin/com/spendoo/designsystem/components/achievementShape.kt
@@ -36,7 +36,7 @@ fun AchievementShape(
     modifier: Modifier = Modifier,
     shapeHexagon: Shape = HexagonShape(),
     shapeRibbon: Shape = RibbonShape(),
-    level: Int
+    level: Long
 ) {
     val backgroundColor = if (isUnlocked) {
         if (Theme.isDarkTheme) {
